requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
Deprecated==1.2.13
Django==4.0
djangorestframework==3.13.1
filelock==3.4.0
flake8==4.0.1
idna==3.3
mccabe==0.6.1
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2021.3
redis==4.0.2
requests==2.26.0
requests-file==1.5.1
six==1.16.0
sqlparse==0.4.2
tldextract==3.1.2
tzdata==2021.5
urllib3==1.26.7
wrapt==1.13.3
